Change txpool demotions type to uint64

+++ txpool/account.go
@@ -150,7 +150,7 @@ type account struct {
 	init               sync.Once
 	enqueued, promoted *accountQueue
 	nextNonce          uint64
-	demotions          uint
+	demotions          uint64
 }
 
 // getNonce returns the next expected nonce for this account.
